ci: Fix test database cleanup (no-changelog) (#6188)
